+++ Tests/Editor/AllocationTests.cs
@@ -110,8 +110,7 @@ class ParamsArrayAllocation
             var issue = issues.First();
 
             Assert.AreNotEqual(16707566, issue.line);
-            Assert.AreEqual("'Closure allocation in ClosureAllocation.System.Void ClosureAllocation::Dummy()'", issue.description);
-            Assert.AreEqual("ClosureAllocation.Dummy", issue.name);
+            Assert.AreEqual("Closure allocation in ClosureAllocation.Dummy", issue.description);
             Assert.AreEqual(IssueCategory.Code, issue.category);
         }